+++ executor_test.go
@@ -41,8 +41,8 @@ func TestExecutor(t *testing.T) {
 			AssertEqual(e.Commands.GetMaxVoltageDetected, true)
 			AssertEqual(e.Commands.GetMinVoltageDetected, true)
 			AssertEqual(e.Commands.GetStopChargeUnderVoltage, true)
-			AssertEqual(e.Commands.SetStopChargeUnderVoltageOn, true)
-			AssertEqual(e.Commands.SetStopChargeUnderVoltageOff, true)
+			AssertEqual(e.Commands.SetStopChargeUnderVoltageOn, false)
+			AssertEqual(e.Commands.SetStopChargeUnderVoltageOff, false)
 			AssertEqual(e.Commands.GetRealTimeVoltage, true)
 			AssertEqual(e.Commands.GetLowVoltage, true)
 			AssertEqual(e.Commands.SetMaxVoltage, float32(3.6))
